requirements.txt
Flask==2.2.0  # 與新版 Werkzeug 兼容
Werkzeug==2.2.0  # 適合 Flask 2.2.0
gunicorn==20.1.0  # 用於啟動 Flask 的 WSGI 伺服器
line-bot-sdk==2.4.3  # 用於 LINE Bot 的 SDK
requests==2.31.0  # 用於 HTTP 請求的庫
groq
pandas
